#255 - Implemented placeholder logic for WV to backend communication
